Crate array_bytes

Enums

Error

The main error of this crate.

Traits

TryFromHex

Simple and safe T/Hex conversions that may fail in a controlled way under some circumstances.

Functions

bytes2hex

Bytes to Hex.

hex2array

Just like hex2bytes but to a fixed length array.

hex2array_unchecked

Just like hex2array but without the checking.

hex2bytes

Hex to Bytes.

hex2bytes_unchecked

Just like hex2bytes but without checking.

hex_bytes2hex_str

Convert hex bytes to hex string.

hex_bytes2hex_str_unchecked⚠

Just like hex_bytes2hex_str but without the checking.

hex_into

Try to convert Hex to T directly, where T: From<Vec<u8>>.

hex_into_unchecked

Just like hex_into but without the checking.

hex_n_into

Try to convert Hex to T directly, where T: From<[u8; N]>.

hex_n_into_unchecked

Just like hex_n_into but without the checking.

slice2array

&[T] to [T; N].

slice2array_unchecked

Just like slice2array but without the checking.

slice_n_into

Convert &[T] to a type directly.

slice_n_into_unchecked

Just like slice_n_into but without the checking.

vec2array

Vec<T> to [T; M].

vec2array_unchecked

Just like vec2array but without the checking.

vec_n_into

Convert Vec<T> to a type directly.

vec_n_into_unchecked

Just like vec_n_into but without the checking.

Type Definitions

ArrayBytesResult

The generic main result of crate array-bytes.

Bytes

Alias for Vec<u8>.

Hex

Alias for String.