Enum array_bytes::Error

pub enum Error {
    InvalidLength,
    InvalidCharacter {
        character: char,
        index: usize,
    },
    MismatchedLength {
        expect: usize,
    },
    Utf8Error(Utf8Error),
    ParseIntError(ParseIntError),
}

The main error of array-bytes.

Variants

InvalidLength

The length must not be odd.

InvalidCharacter

Fields

character: char

The invalid character.

index: usize

The invalid character’s index.

Found the invalid character at index.

MismatchedLength

Fields

expect: usize

Expected length.

The data can not fit the array/slice length well.

Utf8Error(Utf8Error)

Failed to parse the hex number from hex string.

ParseIntError(ParseIntError)

Failed to parse the hex number from hex string.